pyproject.toml
[project]
name = "ig-gmail-mcp"
version = "0.1.0"
description = "Gmail MCP Server"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastmcp>=2.7.1",
    "google-api-python-client>=2.172.0",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.2",
    "mangum>=0.19.0",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.1.0",
]
